consolidate test type helpers into type-helpers.js
