+++ src/application/common/material-aggregation.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
-import { assertPresent } from '#test/assert-present.js'
+import { assertPresent } from '#test/type-helpers.js'
 import { Decimal128 } from 'mongodb'
 import {
   buildEffectiveMaterialStages,
