+++ src/common/helpers/fail-action.test.js
@@ -2,7 +2,7 @@ import Boom from '@hapi/boom'
 import { StatusCodes } from 'http-status-codes'
 import Joi from 'joi'
 
-import { assertPresent } from '#test/assert-present.js'
+import { assertPresent } from '#test/type-helpers.js'
 import { failAction } from './fail-action.js'
 import { config } from '#root/config.js'
