+++ src/common/helpers/formatters.test.js
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { invalidArg } from '#test/invalid-arg.js'
+import { invalidArg } from '#test/type-helpers.js'
 import {
   capitalize,
   formatAddress,
