+++ src/domain/summary-logs/markers.test.js
@@ -1,4 +1,4 @@
-import { invalidArg } from '#test/invalid-arg.js'
+import { invalidArg } from '#test/type-helpers.js'
 import {
   EPR_PREFIX,
   META_PREFIX,
