+++ src/domain/summary-logs/table-schemas/reprocessor-output/validators/uk-packaging-weight-proportion-validator.test.js
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { assertPresent } from '#test/assert-present.js'
+import { assertPresent } from '#test/type-helpers.js'
 import Joi from 'joi'
 import {
   validateUkPackagingWeightProportion,
