+++ src/reports/domain/merge-reporting-periods.test.js
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { assertPresent } from '#test/assert-present.js'
+import { assertPresent } from '#test/type-helpers.js'
 import { mergeReportingPeriods } from './merge-reporting-periods.js'
 
 /**
